Some update requests from Woody: Woody, I updated the stations as you suggested. Also fixed it so it reads T-SUM 200
Also, regarding the precip... it just uses data that is available on the website and i noticed that sometimes it lags a day or two so i added a bullet point that says "Using data as of: " to indicate the date that the data was harvested from.... Perhaps the rainfall you were expecting was fairly recent and hadn't been recorded yet by wunderground?  


John


On Fri, Nov 22, 2013 at 6:58 PM, Woody Lane <woody@woodylane.com> wrote:
Hi John,

Thanks so much for doing this! Looks very good. Here are some alternative weather stations and also a couple of notes or suggestions:

For Mahlon Sweet Airport -- your Junction City station is fine. 
     Just eliminate the choice for Mahlen Sweet Airport.

Coos Bay -- replace with KORCOOSB8
  or if this doesn't work, then use "North Bend"  KORNORTH10
          (end two characters are numbers "1" and "zero")

Port Orford -- replace with KORPORTO2
     (end two characters are the letter "O" and the number "2")

Roseburg -- your Garden Valley station is just fine. This will suffice for the station closest to Roseburg.
    eliminate "KRBG  Roseburg  (downtown airport)"


-  -  -  -  -  -  -  -  -  -  -  -  -  -  

Other suggestions/notes:
1.  Could you change the text word "Tsum200" to "T-SUM 200" in the title and also the results. If the hyphen and blank space cause html problems, then simply capitalize the entire thing as "TSUM200" in all places.

2.  I noticed that some of the sites showed precip for the last 7 days as 0.0. Yet I know that we received more than one inch of rain during that time. Examples: stations of Myrtle Creek, Garden Valley, and Junction City. There may be others also. Do you have any idea what gives?  

Anyway, thanks so much for working on this! It will be incredibly useful, and all the group members will be quite pleased and impressed.

Best regards.